Preparation for gluon v2018.1; Set vxlan to false
In non multi domain setups, vxlan is mandatory.

Release notes cites:
--- snip ---
In single domain setups, the new mesh.vxlan option is mandatory. It
should be set to true in new meshes; existing setups should set it to
false to retain compatibility with older versions of Gluon.
--- snap ---

It seems that we do have to wait until all Nodes are upgraded to
v2018.1, so we set it to false by now.